feat: :sparkles: computedプロパティの型設定
