+++ 05_typescript_v3/src/main.ts
@@ -1,6 +1,7 @@
-import { createApp } from 'vue'
-import App from './App.vue'
+import { createApp } from 'vue';
+// import App from './App/App01.vue'
+import App from './App.vue';
 
-import './assets/main.css'
+import './assets/main.css';
 
-createApp(App).mount('#app')
+createApp(App).mount('#app');
